divided all classes within processor into separate files and adjusted the imports for each file as appropriate
